+++ packages/auth-apple/client/helpers/onLogin.android.js
@@ -5,16 +5,14 @@ import axios from 'axios'
 import { BASE_URL } from '@env'
 import { CALLBACK_NATIVE_URL } from '../../isomorphic'
 
-export default async function onLogin ({ clientId, testBaseUrl }) {
-  const baseUrl = testBaseUrl || BASE_URL
-
+export default async function onLogin ({ baseUrl = BASE_URL, clientId, testBaseUrl }) {
   try {
     const rawNonce = $root.id()
     const state = $root.id()
 
     appleAuthAndroid.configure({
       clientId,
-      redirectUri: baseUrl + CALLBACK_NATIVE_URL,
+      redirectUri: (testBaseUrl || baseUrl) + CALLBACK_NATIVE_URL,
       responseType: appleAuthAndroid.ResponseType.ALL,
       scope: appleAuthAndroid.Scope.ALL,
       nonce: rawNonce,