+++ packages/auth-linkedin/client/components/AuthButton/index.js
@@ -9,8 +9,7 @@ import { faLinkedinIn } from '@fortawesome/free-brands-svg-icons'
 import { LINKEDIN_WEB_LOGIN_URL } from '../../../isomorphic'
 import './index.styl'
 
-function AuthButton ({ label }) {
-  const baseUrl = BASE_URL
+function AuthButton ({ baseUrl, label }) {
   const [authConfig] = useSession('auth')
   const [, $showModal] = useValue(false)
 
@@ -49,11 +48,13 @@ function AuthButton ({ label }) {
 }
 
 AuthButton.defaultProps = {
-  label: 'LinkedIn'
+  label: 'LinkedIn',
+  baseUrl: BASE_URL
 }
 
 AuthButton.propTypes = {
-  label: PropTypes.string.isRequired
+  label: PropTypes.string.isRequired,
+  baseUrl: PropTypes.string.isRequired
 }
 
 export default observer(AuthButton)