+++ packages/auth-local/client/helpers/useAuthHelper.js
@@ -1,11 +1,9 @@
-import { useState, useEffect } from 'react'
+import { useState } from 'react'
+import { BASE_URL } from '@env'
 import AuthHelper from './AuthHelper'
 
-function useAuthHelper () {
-  const [authHelper, setAuthHelper] = useState()
-  useEffect(() => {
-    setAuthHelper(new AuthHelper())
-  }, [])
+function useAuthHelper (baseUrl = BASE_URL) {
+  const [authHelper] = useState(new AuthHelper(baseUrl))
   return authHelper
 }