+++ packages/auth/client/components/index.js
@@ -1,5 +1,6 @@
 export { default as LogoutButton } from './LogoutButton'
 export { default as AuthForm } from './AuthForm'
+export { default as AuthModal } from './AuthModal'
 export { default as SuccessRedirect } from './SuccessRedirect'
 export { default as Logo } from './Logo'
 export { default as OrDivider } from './OrDivider'