+++ packages/ui/components/forms/Multiselect/multiselect.js
@@ -3,7 +3,6 @@ import { ScrollView } from 'react-native'
 import { observer } from 'startupjs'
 import PropTypes from 'prop-types'
 import Drawer from './../../popups/Drawer'
-
 import MultiselectInput from './input'
 import styles from './index.styl'